fix(fx-lib): harden SH/BI decoders; fuzz harness batch 3 (#118) (#394)
* fix(fx-lib): harden SH/BI decoders; fuzz harness batch 3 (#118)

Add libFuzzer harnesses for the ten game-data parsers (brf, mission, sh, t2,
plt, hud, inf, lay, sms, bi) — the epic #49 formats and the new write paths.
Round-trip codecs assert a serialize fixed point (plt asserts byte-identity of
its verbatim round-trip); read-only parsers touch every decoded field under
ASan. Each harness ships a coverage-minimized seed corpus.

The harnesses immediately surfaced six decoder memory-safety bugs, all fixed
here with regression coverage:

- sh find_code_section: `pe_off + 24 + 2` and `raw_ptr + raw_sz` bounds checks
  computed in uint32_t — a huge PE offset or section size wrapped past the
  guard, giving a wild header pointer (SEGV) and a bogus multi-GB code section
  (malloc(~4 GB) OOM in the visited map). Now size_t math.
- bi parse_bi_sections / build_import_map: the same uint32_t-wrap on `pe_off`
  and on the CODE/.idata `raw_ptr + raw_sz` checks (OOB reads), plus an import
  name walk that dereferenced before the bound check (`*np` before
  `np - idata < idata_sz`), reading one byte past the section. Now size_t math
  and bound-before-deref.

Regressions land as explicit unit tests (tests/test_sh.cpp, new
tests/test_bi.cpp) and as crash reproducers in each corpus, which the
fuzz-smoke job replays on every run. Every codec spec's `codec.fuzz` now claims
its harness; STATUS matrix regenerated.

Closes #118

* test(fx-lib): BI disassembly dump snapshots — close #114 (epic #50)

Add the dump-snapshot half of the ai/bi/pe suite (#114): hand-crafted opcode
streams disassemble to a fixed listing, covering a value expression
(push/load/store/arith/eval/end) and forward-jump label annotation, via the
shared make_le harness. With the AI 9-stock round-trip census (test_ai.cpp)
and the PE loader error paths (test_pe.cpp) already present, this closes #114 —
the last open sub-issue of epic #50 (Test & fixture completeness).

Closes #114

+++ lib/src/sh.cpp
@@ -25,7 +25,7 @@ static CodeSection find_code_section(const uint8_t* data, size_t size) {
     if (size < 0x40 || data[0] != 'M' || data[1] != 'Z')
         return {nullptr, 0};
     uint32_t pe_off = u32le(data + 0x3C);
-    if (pe_off + 24 + 2 > size)
+    if ((size_t)pe_off + 24 + 2 > size)  // size_t math: a huge pe_off must not wrap
         return {nullptr, 0};
     const uint8_t* pe = data + pe_off;
     if (pe[0] != 'P' || pe[2] != 0 || pe[3] != 0)
@@ -33,14 +33,14 @@ static CodeSection find_code_section(const uint8_t* data, size_t size) {
     // COFF header at pe+4
     uint16_t num_sec    = u16le(pe + 6);
     uint16_t opt_hdr_sz = u16le(pe + 20);
-    uint32_t sec_table  = pe_off + 24 + opt_hdr_sz;
+    size_t   sec_table  = (size_t)pe_off + 24 + opt_hdr_sz;
     for (uint16_t i = 0; i < num_sec; ++i) {
-        uint32_t sec_off = sec_table + (uint32_t)i * 40;
+        size_t sec_off = sec_table + (size_t)i * 40;  // size_t: no 32-bit wrap
         if (sec_off + 40 > size) break;
         const uint8_t* sec = data + sec_off;
         uint32_t raw_sz  = u32le(sec + 16);
         uint32_t raw_ptr = u32le(sec + 20);
-        if (raw_sz > 0 && raw_ptr + raw_sz <= size)
+        if (raw_sz > 0 && (size_t)raw_ptr + raw_sz <= size)  // no wrap on a huge raw_sz
             return {data + raw_ptr, raw_sz};
     }
     return {nullptr, 0};
